[CCUBE-1522][JY] Rename schema to yup-schema, Create reusable schema helpers.
