+++ src/components/element-editor/prefill/prefill.tsx
@@ -2,7 +2,7 @@ import { Text } from "@lifesg/react-design-system/text";
 import { useFieldArray, useFormContext } from "react-hook-form";
 import { MultiEntry } from "src/components/common";
 import { useBuilder } from "src/context-providers";
-import { SchemaHelper, TFormFieldValues } from "src/schemas";
+import { SchemaHelper, TFormFieldValues } from "src/yup-schemas";
 import * as Yup from "yup";
 import { PrefillChild } from "./prefill-child";
 export const Prefill = () => {
