--- src/components/element-editor/types.ts
@@ -1,6 +1,0 @@
-import { IBaseFieldAttributes } from "src/context-providers/builder";
-
-export interface IFieldValues extends Omit<IBaseFieldAttributes, "internalId"> {
-    placeholder?: string;
-    defaultValue?: string;
-}
