+++ src/yup-schemas/base-helper.ts
@@ -6,8 +6,6 @@ import {
 import * as yup from "yup";
 
 const ID_REGEX = /^[a-z]+(?:[A-Z0-9][a-z0-9]*)*(?:[-_][a-z0-9]+)*$/gm;
-export const PREFILL_ACTIONID_REGEX = /^[a-zA-Z0-9_-]+$/;
-export const PREFILL_PATH_REGEX = /^[a-zA-Z0-9._-]+$/;
 
 export namespace BaseSchemaHelper {
     export const getBaseSchema = (
